backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
pyyaml>=6.0.1
aiofiles>=23.2.1
